ci: simplify type tests
When following the fp-ts de facto standard of a forked version of dtslint, I keep running into the
same problem: different versions of TypeScript produce different types, leading to failed string
matching (such as in https://github.com/gcanti/fp-ts/commit/d5c5935). Since it relies on code
comments, it's also easy to break expectations without realising.

This change switches to using the expect-type library (https://github.com/mmkal/ts) instead. It's
simple and runs as part of the standard tsc typecheck.

I considered using tsd (https://github.com/SamVerschueren/tsd), but it has to run on the compiled
type file, so it is more complicated to use (https://github.com/SamVerschueren/tsd/issues/132).

This change does mean that it only tests it against the version of TypeScript defined in
package.json, but it's possible in the future to install different versions in separate CI jobs.